pyproject.toml
[tool.poetry]
name = "sparrow"
version = "0.1.0"
description = "A tool for automating Helm Chart management via PR"
authors = ["atmask <benjaminajm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
marshmallow = "^3.21.2"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
thefuzz = "^0.22.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
gitpython = "^3.1.43"
pyyaml = "^6.0.2"
azure-identity = "^1.19.0"
azure-mgmt-containerservice = "^32.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
